fix(ui): reset lazy attachment previews
